Moved all data class and enum into GoogleMapDataHelper
