+++ app/src/main/java/com/github/sdpteam15/polyevents/view/activity/MainActivity.kt
@@ -10,10 +10,9 @@ import androidx.fragment.app.Fragment
 import com.github.sdpteam15.polyevents.R
 import com.github.sdpteam15.polyevents.Settings
 import com.github.sdpteam15.polyevents.helper.HelperFunctions
+import com.github.sdpteam15.polyevents.model.map.MapsFragmentMod
 import com.github.sdpteam15.polyevents.model.database.remote.Database.currentDatabase
-import com.github.sdpteam15.polyevents.model.entity.Rating
 import com.github.sdpteam15.polyevents.model.entity.UserEntity
-import com.github.sdpteam15.polyevents.model.observable.Observable
 import com.github.sdpteam15.polyevents.view.fragments.*
 import com.github.sdpteam15.polyevents.view.service.TimerService
 import com.google.android.material.bottomnavigation.BottomNavigationView
@@ -30,7 +29,7 @@ class MainActivity : AppCompatActivity() {
                     mapFragment = HashMap()
                     mapFragment!![R.id.id_fragment_admin_hub] = AdminHubFragment()
                     mapFragment!![R.id.ic_home] = HomeFragment()
-                    mapFragment!![R.id.ic_map] = MapsFragment(MapsFragment.MapsFragmentMod.Visitor)
+                    mapFragment!![R.id.ic_map] = MapsFragment(MapsFragmentMod.Visitor)
                     mapFragment!![R.id.ic_list] = EventListFragment()
                     mapFragment!![R.id.ic_login] = LoginFragment()
                     mapFragment!![R.id.ic_more] = MoreFragment()
